E2E Test Improvements
