+++ test/e2e/tests/general-settings-language.spec.ts
@@ -1,0 +1,61 @@
+import { test, expect } from '@playwright/test';
+import { navigateToAppointmentAndSignIn, getUserSettingsFromLocalStore, setDefaultUserSettingsLocalStore } from '../utils/utils';
+import { SettingsPage } from '../pages/settings-page';
+import { DashboardPage } from '../pages/dashboard-page';
+
+import {
+  PLAYWRIGHT_TAG_E2E_SUITE,
+  PLAYWRIGHT_TAG_PROD_NIGHTLY,
+  PLAYWRIGHT_TAG_PROD_NIGHTLY_MOBILE,
+  APPT_LANGUAGE_SETTING_DE,
+  APPT_LANGUAGE_SETTING_EN,
+  TIMEOUT_3_SECONDS,
+  TIMEOUT_30_SECONDS,
+  APPT_BROWSER_STORE_LANGUAGE_EN,
+  APPT_BROWSER_STORE_LANGUAGE_DE,
+ } from '../const/constants';
+
+let settingsPage: SettingsPage;
+let dashboardPage: DashboardPage;
+
+test.describe('general settings - language', {
+  tag: [PLAYWRIGHT_TAG_E2E_SUITE, PLAYWRIGHT_TAG_PROD_NIGHTLY, PLAYWRIGHT_TAG_PROD_NIGHTLY_MOBILE],
+}, () => {
+  test.beforeEach(async ({ page }) => {
+    // navigate to and sign into appointment
+    await navigateToAppointmentAndSignIn(page);
+    settingsPage = new SettingsPage(page);
+    dashboardPage = new DashboardPage(page);
+
+    // navigate to the general settings page
+    await settingsPage.gotoGeneralSettingsPage();
+
+    // ensure our settings are set to what the tests expect as default (in case a
+    // previous test run failed and left the settings in an incorrect state)
+    await setDefaultUserSettingsLocalStore(page);
+    await page.reload();
+    await page.waitForTimeout(TIMEOUT_3_SECONDS);
+  });
+
+  test('able to change language', async ({ page }) => {
+    // change language setting to DE and verify; we use expect.soft here because if the expect fails
+    // we still want the test to continue so that the test will change the setting back to original value
+    // (expect.soft will still mark the test case failed if the expect fails, but won't stop the test)
+    await settingsPage.changeLanguageSetting(APPT_LANGUAGE_SETTING_DE);
+    await expect.soft(settingsPage.settingsHeaderDE).toBeVisible({ timeout: TIMEOUT_30_SECONDS });
+    await expect.soft(settingsPage.generalSettingsHeaderDE).toBeVisible();
+
+    // verify setting saved in browser local storage
+    let localStore = await getUserSettingsFromLocalStore(page);
+    expect(localStore['language']).toBe(APPT_BROWSER_STORE_LANGUAGE_DE);
+
+    // change language settings back to EN and verify
+    await settingsPage.changeLanguageSetting(APPT_LANGUAGE_SETTING_EN);
+    await expect(settingsPage.settingsHeaderEN).toBeVisible({ timeout: TIMEOUT_30_SECONDS });
+    await expect(settingsPage.generalSettingsHeaderEN).toBeVisible();
+
+    // verify setting saved in browser local storage
+    localStore = await getUserSettingsFromLocalStore(page);
+    expect(localStore['language']).toBe(APPT_BROWSER_STORE_LANGUAGE_EN);
+  });
+});
