+++ test/e2e/tests/general-settings-theme.spec.ts
@@ -1,0 +1,56 @@
+import { test, expect } from '@playwright/test';
+import { navigateToAppointmentAndSignIn, getUserSettingsFromLocalStore, setDefaultUserSettingsLocalStore } from '../utils/utils';
+import { SettingsPage } from '../pages/settings-page';
+import { DashboardPage } from '../pages/dashboard-page';
+
+import {
+  PLAYWRIGHT_TAG_E2E_SUITE,
+  PLAYWRIGHT_TAG_PROD_NIGHTLY,
+  PLAYWRIGHT_TAG_PROD_NIGHTLY_MOBILE,
+  APPT_THEME_SETTING_DARK,
+  APPT_THEME_SETTING_LIGHT,
+  TIMEOUT_3_SECONDS,
+  APPT_BROWSER_STORE_THEME_LIGHT,
+  APPT_BROWSER_STORE_THEME_DARK,
+ } from '../const/constants';
+
+let settingsPage: SettingsPage;
+let dashboardPage: DashboardPage;
+
+test.describe('general settings - theme', {
+  tag: [PLAYWRIGHT_TAG_E2E_SUITE, PLAYWRIGHT_TAG_PROD_NIGHTLY, PLAYWRIGHT_TAG_PROD_NIGHTLY_MOBILE],
+}, () => {
+  test.beforeEach(async ({ page }) => {
+    // navigate to and sign into appointment
+    await navigateToAppointmentAndSignIn(page);
+    settingsPage = new SettingsPage(page);
+    dashboardPage = new DashboardPage(page);
+
+    // navigate to the general settings page
+    await settingsPage.gotoGeneralSettingsPage();
+
+    // ensure our settings are set to what the tests expect as default (in case a
+    // previous test run failed and left the settings in an incorrect state)
+    await setDefaultUserSettingsLocalStore(page);
+    await page.reload();
+    await page.waitForTimeout(TIMEOUT_3_SECONDS);
+  });
+
+  test('able to change theme', async ({ page }) => {
+    // change theme setting to dark mode and verify
+    await settingsPage.changeThemeSetting(APPT_THEME_SETTING_DARK);
+    expect.soft(await settingsPage.isDarkModeEnabled(page)).toBeTruthy();
+
+    // verify setting saved in browser local storage
+    let localStore = await getUserSettingsFromLocalStore(page);
+    expect(localStore['colourScheme']).toBe(APPT_BROWSER_STORE_THEME_DARK);
+
+    // change theme setting back to light mode and verify
+    await settingsPage.changeThemeSetting(APPT_THEME_SETTING_LIGHT);
+    expect(await settingsPage.isDarkModeEnabled(page)).toBeFalsy();
+  
+    // verify setting saved in browser local storage
+    localStore = await getUserSettingsFromLocalStore(page);
+    expect(localStore['colourScheme']).toBe(APPT_BROWSER_STORE_THEME_LIGHT);
+  });
+});
